+++ changelog.md
@@ -1,5 +1,6 @@
 ## 0.3.2
 - add `IEnumerable<Option<T>>.ValuesIntoOrError`
+- add `IEnumerable<T>.TryFirst(predicate)`
 - mark obsolete `OptionsMarshall.TryGetValue` and `TryGetError` as errors (will be removed after 0.3.2)
 - updated TUnit to 1.17.11